# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para PC4 (item-based, Mongo, Redis)",
        "title": "NodosML Movie Recommender API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "credenciales",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Crea un usuario nuevo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "datos",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "tags": [
                    "health"
                ],
                "summary": "Healthcheck",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/me/ratings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ratings"
                ],
                "summary": "Listar MIS ratings",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ratings"
                ],
                "summary": "Crear/actualizar mis ratings",
                "parameters": [
                    {
                        "description": "rating",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ratingRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/me/recommendations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommend"
                ],
                "summary": "Mis recomendaciones",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cantidad de recomendaciones (máx 50)",
                        "name": "k",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "si true, ignora cache Redis",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RecItem"
                            }
                        }
                    }
                }
            }
        },
        "/movies/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Buscar / listar películas (paginado)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "búsqueda por título",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filtrar por género",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "año desde",
                        "name": "year_from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "año hasta",
                        "name": "year_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "límite",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MovieDoc"
                            }
                        }
                    }
                }
            }
        },
        "/movies/top": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Top películas (popularidad o rating)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "popular|rating (default: popular)",
                        "name": "metric",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "límite (default: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MovieDoc"
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "movieId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MovieDoc"
                        }
                    }
                }
            }
        },
        "/users/{id}/ratings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ratings"
                ],
                "summary": "Listar ratings de un usuario (ADMIN)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ratings"
                ],
                "summary": "Crear/actualizar rating (ADMIN)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "rating",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ratingRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/users/{id}/recommendations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommend"
                ],
                "summary": "Recomendaciones para un usuario (ADMIN)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "cantidad de recomendaciones (máx 50)",
                        "name": "k",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "si true, ignora cache Redis",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RecItem"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}/update": {
            "put": {
                "description": "Actualiza los datos de un usuario existente (email, role, password). Todos los campos son opcionales.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Actualizar usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "datos a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}/ws/recommendations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommend"
                ],
                "summary": "Recomendaciones en tiempo real (WebSocket, ADMIN)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "cantidad de recomendaciones (máx 50)",
                        "name": "k",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "si true, ignora cache Redis",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.loginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.ratingRequest": {
            "type": "object",
            "properties": {
                "movieId": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "handler.registerRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "handler.updateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "models.CastMember": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                }
            }
        },
        "models.ExternalData": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "integer"
                },
                "cast": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CastMember"
                    }
                },
                "director": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "posterUrl": {
                    "type": "string"
                },
                "revenue": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "integer"
                },
                "tmdbFetched": {
                    "type": "boolean"
                }
            }
        },
        "models.GenomeTag": {
            "type": "object",
            "properties": {
                "relevance": {
                    "type": "number"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "models.Links": {
            "type": "object",
            "properties": {
                "imdb": {
                    "type": "string"
                },
                "movielens": {
                    "type": "string"
                },
                "tmdb": {
                    "type": "string"
                }
            }
        },
        "models.MovieDoc": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "externalData": {
                    "$ref": "#/definitions/models.ExternalData"
                },
                "genomeTags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GenomeTag"
                    }
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "iIdx": {
                    "type": "integer"
                },
                "links": {
                    "$ref": "#/definitions/models.Links"
                },
                "movieId": {
                    "type": "integer"
                },
                "ratingStats": {
                    "$ref": "#/definitions/models.RatingStats"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userTags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.RatingStats": {
            "type": "object",
            "properties": {
                "average": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "lastRatedAt": {
                    "type": "string"
                }
            }
        },
        "models.RecItem": {
            "type": "object",
            "properties": {
                "movieId": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}